temporary commit

+++ usrun/lib/core/R.dart
@@ -1446,8 +1446,8 @@ class _Styles {
 
 class _ImagePickerDefaults {
   // Sizes are in kilobytes
-  final int maxWidth = 800;
-  final int maxHeight = 600;
+  final double maxWidth = 800;
+  final double maxHeight = 600;
   final int imageQuality = 80;
   final int maxOutputSize = 75;
   final int minInputSize = 500;
